# go/api_gig.go
/*
 * Gigs workforce api
 *
 * This is the Gigs workforce api. Use to search for workforce or search for a gig
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AddGig - Add a new gig to the store
func AddGig(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// DeleteGig - Deletes a gig
func DeleteGig(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// FindGigsByStatus - Finds Gigs by status
func FindGigsByStatus(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// FindGigsByTags - Finds Gigs by tags
// Deprecated
func FindGigsByTags(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetGigById - Find gig by ID
func GetGigById(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// UpdateGigWithForm - Updates a gig in the store with form data
func UpdateGigWithForm(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// UploadFile - uploads an image
func UploadFile(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}