# go/api_store.go
/*
 * Gigs workforce api
 *
 * This is the Gigs workforce api. Use to search for workforce or search for a gig
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// DeleteOrder - Delete purchase order by ID
func DeleteOrder(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetInventory - Returns gig inventories by status
func GetInventory(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetOrderById - Find purchase order by ID
func GetOrderById(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// PlaceOrder - Place an order for a gig
func PlaceOrder(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}