# go/model_category.go
/*
 * Gigs workforce api
 *
 * This is the Gigs workforce api. Use to search for workforce or search for a gig
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Category struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`
}