# go/model_gig.go
/*
 * Gigs workforce api
 *
 * This is the Gigs workforce api. Use to search for workforce or search for a gig
 *
 * API version: 1.0.0
 * Contact: per.von.rosen@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Gig struct {

	Id int64 `json:"id,omitempty"`

	Category Category `json:"category,omitempty"`

	Name string `json:"name"`

	Description []string `json:"description"`

	Measurableoutcome []string `json:"measurableoutcome"`

	Tags []Tag `json:"tags,omitempty"`

	// gig status in the store
	Status string `json:"status,omitempty"`
}